+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 10.1.0 IN PROGRESS
 
+* React v19: refactor away from default props for functional components. Refs UICIRC-1080.
+
 ## [10.0.1](https://github.com/folio-org/ui-circulation/tree/v10.0.1) (2024-12-04)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v10.0.0...v10.0.1)